4. Тесты для стороны прямоугольника. [#159460  [#159460 #412589]]

+++ src/main/java/ru/job4j/condition/SqArea.java
@@ -1,7 +1,7 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(double p, double k) {
+    public static double square(int p, double k) {
         double rsl = ((p * k) / (2 * (k + 1))) * p / (2 * (k + 1));
         return rsl;
     }
